Extract stats into their own file
